chore: rename zindex file
